Moved back nvLink db functions to workflow common
