+++ workflow/pkg/util/common.go
@@ -18,11 +18,16 @@
 package util
 
 import (
+	"context"
 	"slices"
 	"time"
 
+	"github.com/google/uuid"
 	cwutil "github.com/nvidia/bare-metal-manager-rest/common/pkg/util"
+	cdb "github.com/nvidia/bare-metal-manager-rest/db/pkg/db"
 	cdbm "github.com/nvidia/bare-metal-manager-rest/db/pkg/db/model"
+
+	cwssaws "github.com/nvidia/bare-metal-manager-rest/workflow-schema/schema/site-agent/workflows/v1"
 )
 
 var (
@@ -93,4 +98,48 @@ func MachineCapabilitiesEqual(cap1 *cdbm.MachineCapability, cap2 *cdbm.MachineCa
 // IsTimeWithinStaleInventoryThreshold checks if the action time is within the threshold where we could be processing an older inventory
 func IsTimeWithinStaleInventoryThreshold(actionTime time.Time) bool {
 	return time.Since(actionTime) < cwutil.InventoryReceiptInterval+(time.Second*10)
+}
+
+// GetNVLinkLogicalPartitionStatus returns the NVLinkLogicalPartition status and message from Controller NVLinkLogicalPartition state
+func GetNVLinkLogicalPartitionStatus(controllerNVLinkLogicalPartitionTenantState cwssaws.TenantState) (*string, *string) {
+	switch controllerNVLinkLogicalPartitionTenantState {
+	case cwssaws.TenantState_PROVISIONING:
+		return cdb.GetStrPtr(cdbm.NVLinkLogicalPartitionStatusProvisioning), cdb.GetStrPtr("NVLink Logical Partition is being provisioned on Site")
+	case cwssaws.TenantState_CONFIGURING:
+		return cdb.GetStrPtr(cdbm.NVLinkLogicalPartitionStatusConfiguring), cdb.GetStrPtr("NVLink Logical Partition is being configured on Site")
+	case cwssaws.TenantState_READY:
+		return cdb.GetStrPtr(cdbm.NVLinkLogicalPartitionStatusReady), cdb.GetStrPtr("NVLink Logical Partition is ready for use")
+	case cwssaws.TenantState_FAILED:
+		return cdb.GetStrPtr(cdbm.NVLinkLogicalPartitionStatusError), cdb.GetStrPtr("NVLink Logical Partition is in error state")
+	default:
+		return nil, nil
+	}
+}
+
+// UpdateNVLinkLogicalPartitionStatusInDB updates the NVLinkLogicalPartition status in the DB and creates a new StatusDetail
+func UpdateNVLinkLogicalPartitionStatusInDB(ctx context.Context, tx *cdb.Tx, dbSession *cdb.Session, nvlinklogicalpartitionID uuid.UUID, status *string, statusMessage *string) (*cdbm.NVLinkLogicalPartition, *cdbm.StatusDetail, error) {
+	var updatedNVLinkLogicalPartition *cdbm.NVLinkLogicalPartition
+	var err error
+	var newSSD *cdbm.StatusDetail
+	if status != nil {
+		nvlinklogicalpartitionDAO := cdbm.NewNVLinkLogicalPartitionDAO(dbSession)
+		updatedNVLinkLogicalPartition, err = nvlinklogicalpartitionDAO.Update(
+			ctx,
+			tx,
+			cdbm.NVLinkLogicalPartitionUpdateInput{
+				NVLinkLogicalPartitionID: nvlinklogicalpartitionID,
+				Status:                   status,
+			},
+		)
+		if err != nil {
+			return updatedNVLinkLogicalPartition, newSSD, err
+		}
+
+		statusDetailDAO := cdbm.NewStatusDetailDAO(dbSession)
+		newSSD, err = statusDetailDAO.CreateFromParams(ctx, tx, nvlinklogicalpartitionID.String(), *status, statusMessage)
+		if err != nil {
+			return updatedNVLinkLogicalPartition, newSSD, err
+		}
+	}
+	return updatedNVLinkLogicalPartition, newSSD, err
 }
